refactor: constants enum constatns 로 이동 및 주석 제거 (Post)
