+++ src/main/java/com/kt/board/constants/PostDisclosureType.java
@@ -1,0 +1,12 @@
+package com.kt.board.constants;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum PostDisclosureType {
+	PUBLIC("공개"),
+	PRIVATE("비공개");
+
+	private final String description;
+
+}
