+++ src/main/java/com/kt/board/domain/dto/request/PostCreateRequest.java
@@ -1,6 +1,6 @@
 package com.kt.board.domain.dto.request;
 
-import com.kt.board.domain.model.post.PostDisclosureType;
+import com.kt.board.constants.PostDisclosureType;
 
 public record PostCreateRequest(
 	String title,
