+++ src/main/java/com/kt/board/domain/model/post/Post.java
@@ -1,6 +1,6 @@
 package com.kt.board.domain.model.post;
 
-import com.kt.board.domain.dto.request.PostUpdateRequest;
+import com.kt.board.constants.PostDisclosureType;
 import com.kt.board.domain.model.Board;
 import com.kt.board.domain.model.User;
 import com.kt.board.domain.model.common.BaseEntity;
@@ -20,30 +20,27 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Post extends BaseEntity {
 
-	// 게시글 제목
 	private String title;
 
-	// 게시글 내용
 	private String content;
 
-	// 게시글 공개/비공개 여부
 	@Enumerated(EnumType.STRING)
 	private PostDisclosureType disclosureType;
 
-	// Not null
 	@ManyToOne
 	private Board board;
 
-	// Not null
 	@ManyToOne
 	private User user;
 
 	public static Post create(String title, String content, PostDisclosureType disclosureType, Board board, User user) {
 		return new Post(title, content, disclosureType, board, user);
 	}
+
 	public void update(String title, String content, PostDisclosureType disclosureType) {
 		this.title = title;
 		this.content = content;
 		this.disclosureType = disclosureType;
 	}
+
 }
