+++ src/main/java/com/kt/board/service/PostService.java
@@ -2,7 +2,6 @@ package com.kt.board.service;
 
 import com.kt.board.domain.dto.request.PostCreateRequest;
 import com.kt.board.domain.dto.request.PostUpdateRequest;
-import com.kt.board.domain.model.post.PostDisclosureType;
 
 public interface PostService {
